rename: フォルダ名をわかりやすいものに変更
